requirements.txt
torch
torchvision
segmentation-models-pytorch
albumentations
opencv-python-headless
numpy
pillow
matplotlib
tqdm
datasets